+++ fluent-react/examples/fallback-sync/src/index.js
@@ -3,11 +3,11 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import { LocalizationProvider } from 'fluent-react/compat';
 
-import { generateMessages } from './l10n';
+import { generateBundles } from './l10n';
 import App from './App';
 
 ReactDOM.render(
-  <LocalizationProvider messages={generateMessages()}>
+  <LocalizationProvider bundles={generateBundles()}>
     <App />
   </LocalizationProvider>,
   document.getElementById('root')