+++ fluent-react/examples/higher-order/src/l10n.js
@@ -13,7 +13,7 @@ button-show-alert = Click me
   `,
 };
 
-export function* generateMessages(userLocales) {
+export function* generateBundles(userLocales) {
   // Choose locales that are best for the user.
   const currentLocales = negotiateLanguages(
     userLocales,