+++ fluent-react/examples/text-input/src/l10n.js
@@ -17,7 +17,7 @@ type-name =
   `,
 };
 
-export function* generateMessages(userLocales) {
+export function* generateBundles(userLocales) {
   // Choose locales that are best for the user.
   const currentLocales = negotiateLanguages(
     userLocales,