+++ fluent-react/test/provider_change_test.js
@@ -7,13 +7,13 @@ import { LocalizationProvider } from '../src/index';
 suite('LocalizationProvider - changing props', function() {
   test('does not change the ReactLocalization', function() {
     const wrapper = shallow(
-      <LocalizationProvider messages={[]}>
+      <LocalizationProvider bundles={[]}>
         <div />
       </LocalizationProvider>
     );
 
     const oldL10n = wrapper.instance().l10n;
-    wrapper.setProps({ messages: [] });
+    wrapper.setProps({ bundles: [] });
     const newL10n = wrapper.instance().l10n;
 
     assert.equal(oldL10n, newL10n);
@@ -21,27 +21,27 @@ suite('LocalizationProvider - changing props', function() {
 
   test('calls the ReactLocalization\'s setBundles method', function() {
     const wrapper = shallow(
-      <LocalizationProvider messages={[]}>
+      <LocalizationProvider bundles={[]}>
         <div />
       </LocalizationProvider>
     );
 
     const spy = sinon.spy(wrapper.instance().l10n, 'setBundles');
     const newMessages = [];
-    wrapper.setProps({ messages: newMessages });
+    wrapper.setProps({ bundles: newMessages });
     const { args } = spy.getCall(0);
     assert.deepEqual(args, [newMessages]);
   });
 
-  test('changes the ReactLocalization\'s messages bundles', function() {
+  test('changes the ReactLocalization\'s bundles bundles', function() {
     const wrapper = shallow(
-      <LocalizationProvider messages={[]}>
+      <LocalizationProvider bundles={[]}>
         <div />
       </LocalizationProvider>
     );
 
     const oldContexts = wrapper.instance().l10n.bundles;
-    wrapper.setProps({ messages: [] });
+    wrapper.setProps({ bundles: [] });
     const newContexts = wrapper.instance().l10n.bundles;
 
     assert.notEqual(oldContexts, newContexts);